Update index.js
Removed looking for AWS credentials from environment variables if they are not explicitly passed in. the AWS SDK for Node does a lot of magic of managing these credentials. When running this code in an ec2 instance or if we are just relying on default named credentials, this section causes the plugin to fail to authenticate.